Inline downsampling into modular, add qformer_config sub-config, convert to pure functions
- Delete downsampling_granite4_vision.py; move WindowQFormerDownsampler,
  interpolate_downsample, and spatial_offset_downsample into modular
- Replace stateless InterpolateDownsampler/SpatialOffsetDownsampler classes
  with plain functions (items 2 and 4 from reviewer feedback)
- Add config.qformer_config (Blip2QFormerConfig) as a proper sub-config field
  on Granite4VisionConfig following the Blip2Config pattern; remove inline
  Blip2QFormerConfig construction from WindowQFormerDownsampler.__init__ (item 3)
